pyproject.toml
[tool.poetry]
name = "pysoc"
version = "0.1.0"
description = "Social network by Django"
authors = ["Jagi"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Pillow = "^9.1.0"
psycopg2 = "^2.9.3"
djoser = "^2.1.0"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
